requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
openai>=1.0.0

# RAG Toolkit 依赖
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
jieba>=0.42.1
rank-bm25>=0.2.2
numpy>=1.24.0
scipy>=1.10.0
tqdm>=4.65.0
